Profiles: add support for package matchers in build_requires and settings
In order to have more flexibility in deciding to which packages a
setting or build_requires is applied to, introduce the concept of
package matchers. They can be defined in a new section of the profile,
[package_matchers]. Each matcher is a list of package/recipie patterns
(or other matchers) that match a given package/recipie. For example,
one might define the following:

    [package_matchers]
    graphics = libpng/*, libjpeg/*, libtiff/*

A matcher may be used using the $match{NAME} syntax, for example

    [build_requires]
    $match{graphics}: zlib/1.2.11@user/channel

It is now also possible to invert matchers by prepending the pattern
with an exclamation point:

    [build_requires]
    !$match{graphics}: some_package/42@user/channel

Individual patterns may also be prepended:

    [build_requires]
    !libpng/*: libpng/1.6@user/channel

Matchers may also be used in the settings, for example (simplified):

    include(system_compiler)

    [package_matchers]
    toolchain = gcc

    [build_requires]
    !$match{toolchain}: gcc/9.2.0@user/channel

    [settings]
    !$match{toolchain}:compiler.version=9

Matchers are not available for options, as options are very specific
to individual packages.